+++ src/com/jeu/main/GameObject.java
@@ -5,11 +5,11 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {
 
-	protected int x, y;
+	protected float x, y;
 	protected ID id;
-	protected int velocite_X, velocite_Y;
+	protected float velocite_X, velocite_Y;
 	
-	public GameObject(int x, int y, ID id) {
+	public GameObject(float x, float y, ID id) {
 		this.x = x;
 		this.y = y;
 		this.id = id;
@@ -27,11 +27,11 @@ public abstract class GameObject {
 		this.y = y;
 	}
 	
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 	
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	
@@ -51,11 +51,11 @@ public abstract class GameObject {
 		this.velocite_Y = velocite_Y;
 	}
 	
-	public int getVelociteX() {
+	public float getVelociteX() {
 		return velocite_X;
 	}
 
-	public int getVelociteY() {
+	public float getVelociteY() {
 		return velocite_Y;
 	}
